reorg runInPty
